add test

+++ src/DataStructure/WaveletMatrix.hpp
@@ -5,19 +5,19 @@
 #include <cassert>
 template <class T> class WaveletMatrix {
  struct SuccinctIndexableDictionary {
-  size_t len, blocks, zeros;
+  int len, blocks, zeros;
   std::vector<unsigned> bit, sum;
-  SuccinctIndexableDictionary(size_t len): len(len), blocks((len >> 5) + 1), bit(blocks, 0), sum(blocks, 0) {}
+  SuccinctIndexableDictionary(int len): len(len), blocks((len >> 5) + 1), bit(blocks, 0), sum(blocks, 0) {}
   void set(int k) { bit[k >> 5]|= 1U << (k & 31); }
   void build() {
-   for (size_t i= 1; i < blocks; ++i) sum[i]= sum[i - 1] + __builtin_popcount(bit[i - 1]);
+   for (int i= 1; i < blocks; ++i) sum[i]= sum[i - 1] + __builtin_popcount(bit[i - 1]);
    zeros= rank0(len);
   }
   bool operator[](int k) const { return (bit[k >> 5] >> (k & 31)) & 1; }
-  size_t rank(size_t k) const { return (sum[k >> 5] + __builtin_popcount(bit[k >> 5] & ((1U << (k & 31)) - 1))); }
-  size_t rank0(size_t k) const { return k - rank(k); }
+  int rank(int k) const { return (sum[k >> 5] + __builtin_popcount(bit[k >> 5] & ((1U << (k & 31)) - 1))); }
+  int rank0(int k) const { return k - rank(k); }
  };
- size_t len, lg;
+ int len, lg;
  std::vector<SuccinctIndexableDictionary> mat;
  std::vector<T> vec;
 public:
@@ -26,17 +26,17 @@ public:
   std::vector<unsigned> cur(len), nex(len);
   for (int i= len; i--;) cur[i]= std::lower_bound(vec.begin(), vec.end(), v[i]) - vec.begin();
   for (auto h= lg; h--; cur.swap(nex)) {
-   for (size_t i= 0; i < len; ++i)
+   for (int i= 0; i < len; ++i)
     if ((cur[i] >> h) & 1) mat[h].set(i);
    mat[h].build();
    std::array it{nex.begin(), nex.begin() + mat[h].zeros};
-   for (size_t i= 0; i < len; ++i) *it[mat[h][i]]++= cur[i];
+   for (int i= 0; i < len; ++i) *it[mat[h][i]]++= cur[i];
   }
  }
  // k-th(0-indexed) smallest number in v[l,r)
  T kth_smallest(int l, int r, int k) const {
   assert(k < r - l);
-  size_t ret= 0;
+  int ret= 0;
   for (auto h= lg; h--;)
    if (auto l0= mat[h].rank0(l), r0= mat[h].rank0(r); k >= r0 - l0) k-= r0 - l0, ret|= 1 << h, l+= mat[h].zeros - l0, r+= mat[h].zeros - r0;
    else l= l0, r= r0;
@@ -45,16 +45,16 @@ public:
  // k-th(0-indexed) largest number in v[l,r)
  T kth_largest(int l, int r, int k) const { return kth_smallest(l, r, r - l - k - 1); }
  // count i s.t. (l <= i < r) && (v[i] < ub)
- size_t count(int l, int r, T ub) const {
-  size_t x= std::lower_bound(vec.begin(), vec.end(), ub) - vec.begin();
+ int count(int l, int r, T ub) const {
+  int x= std::lower_bound(vec.begin(), vec.end(), ub) - vec.begin();
   if (x >= 1u << lg) return r - l;
   if (x == 0) return 0;
-  size_t ret= 0;
+  int ret= 0;
   for (auto h= lg; h--;)
    if (auto l0= mat[h].rank0(l), r0= mat[h].rank0(r); (x >> h) & 1) ret+= r0 - l0, l+= mat[h].zeros - l0, r+= mat[h].zeros - r0;
    else l= l0, r= r0;
   return ret;
  }
  // count i s.t. (l <= i < r) && (lb <= v[i] < ub)
- size_t count(int l, int r, T lb, T ub) const { return count(l, r, ub) - count(l, r, lb); }
+ int count(int l, int r, T lb, T ub) const { return count(l, r, ub) - count(l, r, lb); }
 };
